Getting rid of some warnings

+++ include/common/GBoundedBufferT.hpp
@@ -237,7 +237,7 @@ public:
 
 #ifdef GEM_COMMON_BENCHMARK_BOUNDED_BUFFER
 		// Update the puts_ vector
-		long currentTime = (boost::posix_time::microsec_clock::local_time() - startTime_).total_microseconds();
+		long currentTime = boost::numeric_cast<long>((boost::posix_time::microsec_clock::local_time() - startTime_).total_microseconds());
 		puts_.push_back(currentTime);
 		entries_.push_back(boost::make_tuple<long, std::size_t>(currentTime, container_.size()));
 #endif /* GEM_COMMON_BENCHMARK_BOUNDED_BUFFER */
@@ -292,7 +292,7 @@ public:
 
 #ifdef GEM_COMMON_BENCHMARK_BOUNDED_BUFFER
 		// Update the puts_ vector
-		long currentTime = (boost::posix_time::microsec_clock::local_time() - startTime_).total_microseconds();
+		long currentTime = boost::numeric_cast<long>((boost::posix_time::microsec_clock::local_time() - startTime_).total_microseconds());
 		puts_.push_back(currentTime);
 		entries_.push_back(boost::make_tuple<long, std::size_t>(currentTime, container_.size()));
 #endif /* GEM_COMMON_BENCHMARK_BOUNDED_BUFFER */
@@ -327,7 +327,7 @@ public:
 
 #ifdef GEM_COMMON_BENCHMARK_BOUNDED_BUFFER
 		// Update the gets_ vector
-		long currentTime = (boost::posix_time::microsec_clock::local_time() - startTime_).total_microseconds();
+		long currentTime = boost::numeric_cast<long>((boost::posix_time::microsec_clock::local_time() - startTime_).total_microseconds());
 		gets_.push_back(currentTime);
 		entries_.push_back(boost::make_tuple<long, std::size_t>(currentTime, container_.size()));
 #endif /* GEM_COMMON_BENCHMARK_BOUNDED_BUFFER */
@@ -403,7 +403,7 @@ public:
 
 #ifdef GEM_COMMON_BENCHMARK_BOUNDED_BUFFER
 		// Update the gets_ vector
-		long currentTime = (boost::posix_time::microsec_clock::local_time() - startTime_).total_microseconds();
+		long currentTime = boost::numeric_cast<long>((boost::posix_time::microsec_clock::local_time() - startTime_).total_microseconds());
 		gets_.push_back(currentTime);
 		entries_.push_back(boost::make_tuple<long, std::size_t>(currentTime, container_.size()));
 #endif /* GEM_COMMON_BENCHMARK_BOUNDED_BUFFER */
@@ -575,7 +575,7 @@ private:
 	 * the root analysis framework (see http://root.cern.ch)
 	 */
 	void emitPutAndGetTimes() {
-      long totalMicroseconds = (endTime_ - startTime_).total_microseconds();
+      long totalMicroseconds = boost::numeric_cast<long>(((endTime_ - startTime_).total_microseconds());
 
       // Create the plot objects
 		boost::shared_ptr<GHistogram1D> gets_ptr(new GHistogram1D(1000, 0, totalMicroseconds));
